Remote commit

+++ src/controllers/UserController.js
@@ -1,15 +1,38 @@
 const bcrypt = require('bcrypt');
 
 module.exports = (app) => {
-  const { User } = app.models;
+  const { User, MensalExpense } = app.models;
 
+  const test = async(req, res) => {
+
+  };
 
   const save = async(req, res) => {
-    let user = {...req.body };
+    let data = {...req.body };
 
     try {
-      const save = await User.create(user);
-      return res.json(save);
+      const check = await User.findOne({
+        where: {
+          login: data.login
+        }
+      });
+
+      if (check) return res.status(400).json({ error: 'Usuário já existente' })
+
+      const user = await User.create(data);
+      const userValid = await User.findOne(data);
+
+      if (!userValid) return res.status(404).json({ error: 'Usuário não encontrado' });
+
+      await MensalExpense.create({ user_id: userValid.id });
+
+      const query = await User.findByPk(user.id, {
+        include: [{
+          association: 'expenses'
+        }]
+      });
+
+      return res.json(query);
     } catch (err) {
       console.log(err)
       return res.status(400).json({ error: "Erro ao salvar usuário" })
@@ -17,11 +40,22 @@ module.exports = (app) => {
   };
 
   const list = async(req, res) => {
-    const user = await User.findAll({
-      //attributes: { exclude: ['password'] }
-    });
 
-    return res.json(user);
+    try {
+      const user = await User.findAll({
+        include: [{
+          association: 'expenses'
+        }],
+        attributes: {
+          exclude: ['password']
+        }
+      });
+
+      return res.json(user);
+    } catch (err) {
+      console.log(err);
+      return res.status(400).json({ error: 'Não foi possível listar os usuários' })
+    }
   };
 
   const search = async(req, res) => {
@@ -29,23 +63,27 @@ module.exports = (app) => {
 
     try {
       const user = User.findByPk(id);
-      res.json(user);
+      return res.json(user);
     } catch (err) {
       console.log(err);
-      res.status(404).json({ error: 'Usuário não encontrado' })
+      return res.status(404).json({ error: 'Usuário não encontrado' })
     }
   };
 
   const remove = async(req, res) => {
     const { user_id: id } = req.params;
     try {
+      const userExistis = await User.findByPk(id);
+      if (!userExistis) {
+        return res.status(404).json({ error: 'Usuário não encontrado' });
+      }
       await User.destroy({
         where: { id }
       });
-      return res.status(200).json({ success: "Sucesso ao remover usuário" })
+      return res.status(200).json({ success: 'Sucesso ao remover usuário' });
     } catch (err) {
       console.log(err);
-      return res.status(400).json({ error: "Erro ao remover usuário" })
+      return res.status(400).json({ error: 'Erro ao remover usuário' });
     };
 
   };
@@ -58,7 +96,7 @@ module.exports = (app) => {
       const user = await User.findByPk(id);
 
       if (!user) {
-        res.status(404).json({ error: 'User not found' });
+        return res.status(404).json({ error: 'User not found' });
       }
 
       const hashPassword = bcrypt.hashSync(data.password, bcrypt.genSaltSync(10));
@@ -67,9 +105,9 @@ module.exports = (app) => {
       const update = await User.update(data, {
         where: { id }
       });
-      res.json(update);
+      return res.json(update);
     } catch (err) {
-      res.status(400).json({ error: "Ocorreu um erro ao atualizar usuário" })
+      return res.status(400).json({ error: "Ocorreu um erro ao atualizar usuário" })
     };
   }
   return {
@@ -77,6 +115,7 @@ module.exports = (app) => {
     list,
     remove,
     search,
-    update
-  }
+    update,
+    //test
+  };
 }
